+++ chat-widget/src/common/telemetry/TelemetryConstants.ts
@@ -294,6 +294,8 @@ export enum TelemetryEvent {
     UXCitationPaneCompleted = "UXCitationPaneCompleted",
     UXLiveChatWidgetStart = "UXLiveChatWidgetStart",
     UXLiveChatWidgetCompleted = "UXLiveChatWidgetCompleted",
+    UXPostChatPaneStarted = "UXPostChatPaneStarted",
+    UXPostChatPaneCompleted = "UXPostChatPaneCompleted",
 
     AppInsightsInitialized = "AppInsightsInitialized",
     AppInsightsInitFailed = "AppInsightsInitFailed",
@@ -461,4 +463,10 @@ export class TelemetryConstants {
     public static mapEventToScenario(eventTypeOrScenarioType: TelemetryEvent): ScenarioType {
         return TelemetryConstants.map(eventTypeOrScenarioType);
     }
+}
+
+export enum ConversationStage {
+    Initialization = "Initialization",
+    CSREngagement = "CSR Engagement",
+    ConversationEnd = "Conversation End"
 }