+++ chat-widget/src/components/livechatwidget/common/startChatErrorHandler.ts
@@ -1,5 +1,5 @@
 import { ChatSDKError, ChatSDKErrorName } from "@microsoft/omnichannel-chat-sdk";
-import { LogLevel, TelemetryEvent } from "../../../common/telemetry/TelemetryConstants";
+import { ConversationStage, LogLevel, TelemetryEvent } from "../../../common/telemetry/TelemetryConstants";
 import { PrepareEndChatDescriptionConstants, WidgetLoadCustomErrorString, WidgetLoadTelemetryMessage } from "../../../common/Constants";
 import { callingStateCleanUp, chatSDKStateCleanUp, closeChatStateCleanUp, endChatStateCleanUp } from "./endChat";
 
@@ -110,7 +110,8 @@ const logWidgetLoadFailed = (ex?: ChatSDKError) => {
         Event: TelemetryEvent.WidgetLoadFailed,
         Description: "Widget load complete with error",
         ExceptionDetails: exDetails,
-        ElapsedTimeInMilliseconds: TelemetryTimers?.WidgetLoadTimer?.milliSecondsElapsed
+        ElapsedTimeInMilliseconds: TelemetryTimers?.WidgetLoadTimer?.milliSecondsElapsed,
+        CustomProperties: { ConversationStage: ConversationStage.Initialization }
     });
 };
 
@@ -155,7 +156,8 @@ const logWidgetLoadCompleteWithError = (ex: ChatSDKError) => {
         Event: TelemetryEvent.WidgetLoadFailed,
         Description: "Widget load complete with error",
         ExceptionDetails: exDetails,
-        ElapsedTimeInMilliseconds: TelemetryTimers?.WidgetLoadTimer?.milliSecondsElapsed
+        ElapsedTimeInMilliseconds: TelemetryTimers?.WidgetLoadTimer?.milliSecondsElapsed,
+        CustomProperties: { ConversationStage: ConversationStage.Initialization }
     });
 };
 
@@ -184,7 +186,8 @@ export const logWidgetLoadWithUnexpectedError = (ex: any) => { // eslint-disable
         Event: TelemetryEvent.WidgetLoadFailed,
         Description: "Widget load with unexpected error",
         ExceptionDetails: exDetails,
-        ElapsedTimeInMilliseconds: TelemetryTimers?.WidgetLoadTimer?.milliSecondsElapsed
+        ElapsedTimeInMilliseconds: TelemetryTimers?.WidgetLoadTimer?.milliSecondsElapsed,
+        CustomProperties: { ConversationStage: ConversationStage.Initialization }
     });
 };